scripts/requirements.txt
psycopg2-binary>=2.9.9
pandas>=2.0.0
schedule>=1.2.0
psutil>=5.9.0
numpy>=1.24.0
requests>=2.28.0